* 完善http2支持

+++ baseio-codec/src/main/java/com/generallycloud/nio/codec/http2/future/Http2HeadersFrame.java
@@ -26,7 +26,7 @@ public interface Http2HeadersFrame extends Http2Frame {
 
 	public abstract int getStreamDependency();
 
-	public abstract int getWeight();
+	public abstract short getWeight();
 	
 	public byte getPadLength();
 }
